moved layout to separate folder, refactore Resource to provider
