+++ examples/ag-demo-ui/src/layout/fresh/index.html
