+++ examples/ag-demo-ui/src/layout/fresh/sidenav/index.js
