+++ examples/ag-demo-ui/src/layout/fresh/sidenav/simple-sidenav.css
