+++ examples/ag-demo-ui/src/layout/fresh/styles/_index.scss
