+++ examples/ag-demo-ui/src/layout/packet/footer/footer.html
