+++ examples/ag-demo-ui/src/layout/views/partials/footer.html
