+++ examples/ag-demo-ui/src/layout/views/partials/main-content.html
