+++ examples/ag-demo-ui/src/layout/views/partials/off-sidebar.html
