+++ examples/ag-demo-ui/src/layout/views/partials/settings.html
