+++ examples/ag-demo-ui/src/layout/views/partials/top-navbar.html
